feat(ts): rename file
