+++ types/global.d.ts
@@ -4,7 +4,7 @@ declare global {
   type BaseColorDesignToken = import('./colors').BaseColorDesignToken;
   type ColorDesignToken = import('./colors').ColorDesignToken;
   type Color = import('./colors').Color;
-  type FontFamily = import('./font-family').FontFamily;
+  type FontFamily = import('./font').FontFamily;
   type ValidationRule = import('./forms').ValidationRule;
   type Span = import('./grid').Span;
   type Icon = import('./icons').Icon;
